Moved unit tests to corresponding files Disabled tests for serialization
