+++ src/prodbg/viewdock/src/rect.rs
@@ -53,3 +53,57 @@ impl Rect {
         (rect_left, rect_right)
     }
 }
+
+#[cfg(test)]
+mod test {
+    use Rect;
+
+    fn check_range(inv: f32, value: f32, delta: f32) -> bool {
+        (inv - value).abs() < delta
+    }
+
+    #[test]
+    fn test_calc_rect_horz_half() {
+        let rects = Rect::split_horizontally(&Rect::new(0.0, 0.0, 1024.0, 1024.0), 0.5);
+
+        assert_eq!(check_range(rects.0.x, 0.0, 0.001), true);
+        assert_eq!(check_range(rects.0.y, 0.0, 0.001), true);
+        assert_eq!(check_range(rects.0.width, 1024.0, 0.001), true);
+        assert_eq!(check_range(rects.0.height, 512.0, 0.001), true);
+
+        assert_eq!(check_range(rects.1.x, 0.0, 0.001), true);
+        assert_eq!(check_range(rects.1.y, 512.0, 0.001), true);
+        assert_eq!(check_range(rects.1.width, 1024.0, 0.001), true);
+        assert_eq!(check_range(rects.1.height, 512.0, 0.001), true);
+    }
+
+    #[test]
+    fn test_calc_rect_horz_25_per() {
+        let rects = Rect::split_horizontally(&Rect::new(0.0, 0.0, 1024.0, 1024.0), 0.25);
+
+        assert_eq!(check_range(rects.0.x, 0.0, 0.001), true);
+        assert_eq!(check_range(rects.0.y, 0.0, 0.001), true);
+        assert_eq!(check_range(rects.0.width, 1024.0, 0.001), true);
+        assert_eq!(check_range(rects.0.height, 256.0, 0.001), true);
+
+        assert_eq!(check_range(rects.1.x, 0.0, 0.001), true);
+        assert_eq!(check_range(rects.1.y, 256.0, 0.001), true);
+        assert_eq!(check_range(rects.1.width, 1024.0, 0.001), true);
+        assert_eq!(check_range(rects.1.height, 768.0, 0.001), true);
+    }
+
+    #[test]
+    fn test_calc_rect_horz_25_per_2() {
+        let rects = Rect::split_horizontally(&Rect::new(16.0, 32.0, 512.0, 1024.0), 0.25);
+
+        assert_eq!(check_range(rects.0.x, 16.0, 0.001), true);
+        assert_eq!(check_range(rects.0.y, 32.0, 0.001), true);
+        assert_eq!(check_range(rects.0.width, 512.0, 0.001), true);
+        assert_eq!(check_range(rects.0.height, 256.0, 0.001), true);
+
+        assert_eq!(check_range(rects.1.x, 16.0, 0.001), true);
+        assert_eq!(check_range(rects.1.y, 288.0, 0.001), true);
+        assert_eq!(check_range(rects.1.width, 512.0, 0.001), true);
+        assert_eq!(check_range(rects.1.height, 768.0, 0.001), true);
+    }
+}
